659 create lyche about us page (#665)
* routs, aboutpage draft and feature to SultenCard

* added small comment

* changed 'LychePage' to 'LycheHomePage'

* made lyche about page

* fixed formating

* removed superfluous style of text container

* fix style sultenCard. added flex-column mixin

* fixed page import in rotes and added to index.ts

* tweeks textItem. fixes dynamic style issue v2
