+++ frontend/src/Components/Page/Page.module.scss
@@ -22,4 +22,3 @@ $side-padding: 1em;
     width: calc(100% - $side-padding * 2);
   }
 }
-
