+++ frontend/src/Components/SultenCard/SultenCard.module.scss
@@ -3,42 +3,70 @@
 @import 'src/mixins';
 
 .container {
-  display: flex;
+  @include flex-row;
   background-color: $black;
+  border-radius: 0.5em;
+  gap: 1em;
+  padding: 1.5em;
   @include for-mobile-down {
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 1em;
+    @include flex-column;
     padding-bottom: 2em;
+  }
+}
+
+.smallcard_container {
+  @include flex-row-center;
+  background-color: $black;
+  border-radius: 0.5em;
+  gap: 1em;
+  padding: 1.5em;
+  @include for-mobile-down {
+    @include flex-column;
   }
 }
 
 .card_image {
   height: 30em;
   object-position: top;
+  object-fit: cover;
   width: 50%;
+  flex: 1;
   display: inline-block;
   border-radius: 0.5em;
-  object-fit: cover;
+
   @include for-mobile-down {
     width: 100%;
     display: block;
+    max-height: 15em;
+  }
+}
+
+.smallcard_image {
+  height: 15em;
+  width: 25%;
+  max-width: 15em;
+  display: inline-block;
+  flex: 1;
+  border-radius: 0.5em;
+  object-position: top;
+  object-fit: cover;
+
+  @include for-mobile-down {
+    width: 100%;
+    max-height: 5em;
+    display: block;
+    object-position: center;
   }
 }
 
 .text_container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+  @include flex-column-center;
+  flex: 1;
   text-align: center;
-  gap: 1em;
-  width: 50%;
-  padding: 0 4rem;
+  max-width: 30vw;
+
   @include for-mobile-down {
-    width: 100%;
-    padding: 0 1rem;
+    max-width: 80vw;
   }
 }
 
@@ -54,4 +82,5 @@
 
 .card_button {
   width: fit-content;
+  margin: 1.5em;
 }
