+++ frontend/src/Pages/LycheAboutPage/index.ts
@@ -1,0 +1,1 @@
+export { LycheAboutPage } from './LycheAboutPage';
