+++ frontend/src/Pages/LycheHomePage/index.ts
@@ -1,0 +1,1 @@
+export { LycheHomePage } from './LycheHomePage';
