--- frontend/src/Pages/LychePage/index.ts
@@ -1,1 +1,0 @@
-export { LychePage } from './LychePage';
