+++ frontend/src/Pages/index.ts
@@ -10,7 +10,8 @@ export { HomePage } from './HomePage';
 export { InformationListPage } from './InformationListPage';
 export { InformationPage } from './InformationPage';
 export { LoginPage } from './LoginPage';
-export { LychePage } from './LychePage';
+export { LycheAboutPage } from './LycheAboutPage';
+export { LycheHomePage } from './LycheHomePage';
 export { NotFoundPage } from './NotFoundPage';
 export { RecruitmentAdmissionFormPage } from './RecruitmentAdmissionFormPage';
 export { RecruitmentPage } from './RecruitmentPage';
