+++ frontend/src/_constants.scss
@@ -64,6 +64,7 @@ $content-padding: calc((100vw - 1200px) / 2);
 
 // Navbar size
 $navbar-height: 60px;
+$sulten-navbar-height: 100px;
 
 // Content size
 $primary-content-wide-padding: 2.5rem;
