+++ frontend/src/_mixins.scss
@@ -172,3 +172,15 @@ Usage:
 @mixin shadow-light {
   box-shadow: 0 0 0.5em 0.1em rgba(0, 0, 0, 0.05);
 }
+
+@mixin flex-column {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+@mixin flex-column-center {
+  @include flex-column;
+  justify-content: center;
+}
