+++ frontend/src/constants/TextItems.ts
@@ -13,4 +13,8 @@ export const TextItem = {
   sulten_menu_text: 'sulten_menu_text',
   sulten_about_us_text: 'sulten_about_us_text',
   sulten_contact_text: 'sulten_contact_text',
+  sulten_about_page_text: 'sulten_about_page_text',
+  sulten_what_is_lyche_text: 'sulten_what_is_lyche_text',
+  sulten_lyche_goal_text: 'sulten_lyche_goal_text',
+  sulten_lyche_about_menu_text: 'sulten_lyche_about_menu_text',
 } as const;
